fix math import in zernike. remove deprecated scipy interp2d

+++ aotools/functions/zernike.py
@@ -1,4 +1,7 @@
+import math
+
 import numpy
+
 from . import circle
 
 def phaseFromZernikes(zCoeffs, size, norm="noll", rot=0):
@@ -90,8 +93,8 @@ def zernikeRadialFunc(n, m, r):
         R += numpy.array(r**(n - 2 * i) * (((-1)**(i)) *
                          math.factorial(n - i)) /
                          (math.factorial(i) *
-                          math.factorial(0.5 * (n + m) - i) *
-                          math.factorial(0.5 * (n - m) - i)),
+                          math.factorial(int(0.5 * (n + m) - i)) *
+                          math.factorial(int(0.5 * (n - m) - i))),
                          dtype='float')
     return R
